Base NPC auto-eating on hunger, not stored kCal
NPCs would only eat if they had less than 95% of their healthy kCals
remaining, however they would still be *hungry*, and would still
complain about this, refuse to train the PC, etc. This would result in
the annoying situation of having to endure a whingy NPC until they learn
to eat for themselves, or spoon-feeding them and having them be
completely player-dependent.

This PR attempts to unify what causes the NPC to complain or not
cooperate (their hunger level) with when they eat, so we're less likely
to have NPCs complain about being hungry with food in their inventory.